added multipool high hashrate notifier
Added multipool high hashrate notifier. It will send an email with
notification if multipool is more than 50%, and give levels based upon
that with 7 being 50% and 1 being 99% or greater.

Known bugs (too tired to fix now) include improper headers being set.

Run as often as you feel you should be notified of multipool being on
the network (I do it once every minute, there are a number of
texts/emails that come during the time they are hashing but I would
prefer the info. you can call it something like every 5-10 minutes and
be fine probably